feat (Real time update)- Chat restructure
- Added ability to send message on enter key
- Add real time updates when person check in and out

+++ customers/views.py
@@ -1,9 +1,11 @@
+from django.core import serializers
 import json
 import random
 from datetime import datetime, timedelta
 
 from django.http import HttpResponse, JsonResponse
 from django.shortcuts import get_object_or_404, redirect, render
+from pkg_resources import safe_extra
 
 from .models import Customer, Teller
 
@@ -26,14 +28,15 @@ def index(request):
 def teller_view(request, pk):
     teller = get_object_or_404(Teller, id=pk)
     if teller:
+        customers = customer_clean(teller.get_customers())
         customers = teller.get_customers()
-        customers = customer_clean(customers)
         next_customer = teller.get_next_customer()
         millisec = datetime.now().isoformat()
+        now = datetime.now().isoformat()
         if next_customer:
-
             millisec = next_customer.expired_date.isoformat()
-        return render(request, "teller.html", context={"teller": teller, "customers": customers, "next_time": millisec})
+
+        return render(request, "teller.html", context={"teller": teller, "customers": customers, "next_time": millisec, "current_time": now})
     return HttpResponse('Not found')
 
 
@@ -49,28 +52,35 @@ def register_customer(request, pk):
 
 def delete_customer(request, pk):
     customer = get_object_or_404(Customer, id=pk)
-    teller = customer.teller
+    if not customer:
+        return HttpResponse({"delete": True})
     customer.delete()
-    return redirect(teller)
+    return HttpResponse({"delete": False})
 
 
 def get_waiting_time(request, pk):
     teller = get_object_or_404(Teller, id=pk)
     if teller:
         now = datetime.now()
-        customers = teller.get_customers()
+        serializer = serializers.serialize(
+            "json", teller.get_customers())
         next_customer = teller.get_next_customer()
         if next_customer:
+            next_customer = {
+                "name": next_customer.name,
+                "code": next_customer.code,
+                "arrived_at": next_customer.arrived_at.isoformat()
+            }
             data = {
-                "customers": customers,
+                "customers": serializer,
                 "next_customer": next_customer,
                 "success": True
             }
-            return JsonResponse(json.dumps(data))
+            return JsonResponse(json.dumps(data), safe=False)
 
         return JsonResponse({
             "success": False,
-            "reason": "No more customerz"
+            "reason": "No more customers"
         })
     return JsonResponse({
         "success": False,
